Convert the view router to an EnvironmentObject
Using an `ObservedObject` that is observed by a `StateObject` can become
confusing when we work with more complex view hierarchies.

An `EnvironmentObject` can be used to share information across all views
of the app. So once we define our view router as an `EnvironmentObject`,
all views will be bound to it without the need of an initialization
chain down the view hierarchy.