Use Swift’s Result type in the networking logic
The `completionHandler` will either have a `data` or `error` set to a
value — it cannot have both or neither because both scenarios do not
make sense. However, `URLSession` does not enforce this constraint!

The `Result` type gives us the either/or behavior we want. It also works
great with non-blocking functions. Furthermore, `Result` lets us return
specific types of errors, which makes it easier to know what went wrong.